backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
attrs==25.3.0
backoff==2.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
clerk-sdk==0.2.2
Deprecated==1.2.18
Django==5.2.4
django-clerk==0.1.15
djangorestframework==3.16.0
dotenv==0.9.9
factory_boy==3.3.3
Faker==37.4.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
promise==2.3
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svix==1.68.0
text-unidecode==1.3
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.2
